backend/requirements.txt
flask
flask-cors
textblob
python-dotenv